requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
Brotli==1.0.9
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
decorator==5.1.1
discord.py==2.3.1
filelock==3.12.2
flake8==6.1.0
flake8-variables-names==0.0.6
frozenlist==1.4.0
idna==3.4
mccabe==0.7.0
multidict==6.0.4
mutagen==1.46.0
pycodestyle==2.11.0
pycparser==2.21
pycryptodomex==3.18.0
pyflakes==3.1.0
PyNaCl==1.5.0
python-dotenv==1.0.0
requests==2.31.0
typing_extensions==4.7.1
urllib3==2.0.4
wavelink==2.6.1
websockets==11.0.3
yarl==1.8.2
